feat(adieu-razzle): moved workbox into build-utils
