+++ packages/falcon-client/build-utils/webpack/tools.js
@@ -87,9 +87,17 @@ function getBuildConfig(buildConfigFileName = 'falcon-client.build.config.js') {
   return config;
 }
 
+function formatBytes(bytes) {
+  if (bytes < 1024) return `${bytes} Bytes`;
+  else if (bytes < 1048576) return `${(bytes / 1024).toFixed(3)} KB`;
+  else if (bytes < 1073741824) return `${(bytes / 1048576).toFixed(3)} MB`;
+  return `${(bytes / 1073741824).toFixed(3)} GB`;
+}
+
 module.exports = {
   getBuildConfig,
   removePreviousBuildAssets,
   webpackCompiler,
-  webpackCompileAsync
+  webpackCompileAsync,
+  formatBytes
 };
